+++ packages/cozy-doctypes/src/Application.js
@@ -1,10 +1,14 @@
 const Document = require('./Document')
 
+const APP_DOCTYPE = 'io.cozy.apps'
+
 class Application extends Document {}
 
 Application.schema = {
-  doctype: 'io.cozy.apps',
+  doctype: APP_DOCTYPE,
   attributes: {}
 }
+
+Application.doctype = APP_DOCTYPE
 
 module.exports = Application