+++ packages/cozy-doctypes/src/banking/BankAccountStats.spec.js
@@ -74,6 +74,29 @@ describe('BankAccountStats::sum', () => {
     })
   })
 
+  it('should return 0 for a missing property', () => {
+    const accountsStats = [
+      {
+        income: 2000,
+        additionalIncome: 400,
+        mortgage: 650,
+        loans: 800,
+        currency: 'EUR'
+      },
+      {
+        income: 1500,
+        additionalIncome: 0,
+        mortgage: 0,
+        loans: 0,
+        currency: 'EUR'
+      }
+    ]
+
+    const sum = BankAccountStats.sum(accountsStats)
+
+    expect(sum.fixedCharges).toBe(0)
+  })
+
   it('should throw an error if the stats currencies are not the same', () => {
     const accountsStats = [
       {