+++ packages/commitlint-config-cozy/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.27](https://github.com/cozy/cozy-libs/compare/commitlint-config-cozy@0.3.26...commitlint-config-cozy@0.3.27) (2019-07-19)
+
+**Note:** Version bump only for package commitlint-config-cozy
+
+
+
+
+
 ## [0.3.26](https://github.com/cozy/cozy-libs/compare/commitlint-config-cozy@0.3.25...commitlint-config-cozy@0.3.26) (2019-07-19)
 
 **Note:** Version bump only for package commitlint-config-cozy